functions.js, event.js and features.js removed
